+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unpublished
 
-- Fix: Only zoom to the user position if the user is currently in the extent of the map.
+- Fix: Only zoom to the user position if the user is currently in the extent of the map. The user position is no longer indicated when it's outside the map's extent.
 
 ## 1.3.0