+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Breaking: As a result of the bundling with `rollup`, the styles of the package need to be imported via `@polar/plugin-geo-location/styles.css`.
 - Feature: The package is now being bundled by `rollup` before being published. This allows for a smaller package size and better compatibility with other packages.
+- Fix: Only zoom to the user position if the user is currently in the extent of the map. The user position is no longer indicated when it's outside the map's extent.
 
 ## 1.3.0