+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unpublished
+## 1.3.1
 
 - Fix: Only zoom to the user position if the user is currently in the extent of the map. The user position is no longer indicated when it's outside the map's extent.
 - Fix: The plugin would fail the map initialization procedure on Safari 15 and prior. This issue has been resolved by disabling the fail-fast feature for browsers not supporting `navigator.permissions.query`.